chore: renamed index files
